backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
openai>=1.50.0
scikit-learn==1.4.0
nltk==3.8.1
joblib==1.3.2
numpy==1.26.3
pandas==2.2.0
python-multipart==0.0.6
